Message API

+++ src/main/java/org/motometer/telegram/bot/api/Chat.java
@@ -13,7 +13,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 public interface Chat {
 
     @JsonProperty("id")
-    int id();
+    long id();
 
     @JsonProperty("type")
     ChatType type();
